Create parent class for Blob*RefProperty classes
Fixes: SE-9561
