+++ src/main/java/sirius/biz/storage/layer2/BlobHardRefProperty.java
@@ -12,14 +12,11 @@ import sirius.biz.storage.Storage;
 import sirius.biz.storage.util.StorageUtils;
 import sirius.db.jdbc.OMA;
 import sirius.db.jdbc.schema.SQLPropertyInfo;
-import sirius.db.jdbc.schema.Table;
-import sirius.db.jdbc.schema.TableColumn;
 import sirius.db.mixing.AccessPath;
 import sirius.db.mixing.BaseEntity;
 import sirius.db.mixing.EntityDescriptor;
 import sirius.db.mixing.Property;
 import sirius.db.mixing.PropertyFactory;
-import sirius.kernel.commons.Value;
 import sirius.kernel.di.std.Part;
 import sirius.kernel.di.std.Register;
 import sirius.kernel.health.Exceptions;
@@ -27,15 +24,12 @@ import sirius.kernel.health.Exceptions;
 import javax.annotation.Nonnull;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
-import java.sql.Types;
 import java.util.function.Consumer;
 
 /**
  * Handles fields of the type {@link BlobHardRef} within an {@link BaseEntity}.
  */
-public class BlobHardRefProperty extends Property implements SQLPropertyInfo {
-
-    protected static final int DEFAULT_KEY_LENGTH = 64;
+public class BlobHardRefProperty extends BlobRefProperty implements SQLPropertyInfo {
 
     /**
      * Factory for generating properties based on their field type
@@ -70,62 +64,6 @@ public class BlobHardRefProperty extends Property implements SQLPropertyInfo {
                                 @Nonnull AccessPath accessPath,
                                 @Nonnull Field field) {
         super(descriptor, accessPath, field);
-    }
-
-    protected BlobHardRef getRef(Object entity) {
-        try {
-            return (BlobHardRef) super.getValueFromField(this.accessPath.apply(entity));
-        } catch (Exception e) {
-            throw Exceptions.handle()
-                            .to(OMA.LOG)
-                            .error(e)
-                            .withSystemErrorMessage(
-                                    "Unable to obtain a reference object from entity ref field ('%s' in '%s'): %s (%s)",
-                                    getName(),
-                                    descriptor.getType().getName())
-                            .handle();
-        }
-    }
-
-    @Override
-    protected Object getValueFromField(Object target) {
-        return getRef(target).getKey();
-    }
-
-    @Override
-    public Object transformValue(Value value) {
-        return value.get();
-    }
-
-    @Override
-    protected Object transformToJDBC(Object object) {
-        return object;
-    }
-
-    @Override
-    protected Object transformFromJDBC(Value object) {
-        return object.get();
-    }
-
-    @Override
-    public void setValue(Object entity, Object object) {
-        this.setValueToField(object, entity);
-    }
-
-    @Override
-    protected void setValueToField(Object value, Object target) {
-        BlobHardRef ref = getRef(target);
-
-        if (value == null || value instanceof Blob) {
-            ref.setBlob((Blob) value);
-        } else {
-            ref.setKey((String) value);
-        }
-    }
-
-    @Override
-    public void contributeToTable(Table table) {
-        table.getColumns().add(new TableColumn(this, Types.CHAR));
     }
 
     @Override
