+++ src/main/java/sirius/biz/storage/layer2/BlobSoftRefProperty.java
@@ -12,13 +12,10 @@ import sirius.biz.storage.Storage;
 import sirius.biz.storage.util.StorageUtils;
 import sirius.db.jdbc.OMA;
 import sirius.db.jdbc.schema.SQLPropertyInfo;
-import sirius.db.jdbc.schema.Table;
-import sirius.db.jdbc.schema.TableColumn;
 import sirius.db.mixing.AccessPath;
 import sirius.db.mixing.EntityDescriptor;
 import sirius.db.mixing.Property;
 import sirius.db.mixing.PropertyFactory;
-import sirius.kernel.commons.Value;
 import sirius.kernel.di.std.Part;
 import sirius.kernel.di.std.Register;
 import sirius.kernel.health.Exceptions;
@@ -26,13 +23,12 @@ import sirius.kernel.health.Exceptions;
 import javax.annotation.Nonnull;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
-import java.sql.Types;
 import java.util.function.Consumer;
 
 /**
  * Handles fields of the type {@link BlobSoftRef} within an {@link sirius.db.mixing.BaseEntity}.
  */
-public class BlobSoftRefProperty extends Property implements SQLPropertyInfo {
+public class BlobSoftRefProperty extends BlobRefProperty implements SQLPropertyInfo {
 
     /**
      * Contains the field length to use if URLs are also supported for this field.
@@ -74,72 +70,17 @@ public class BlobSoftRefProperty extends Property implements SQLPropertyInfo {
         super(descriptor, accessPath, field);
     }
 
-    protected BlobSoftRef getRef(Object entity) {
-        try {
-            return (BlobSoftRef) super.getValueFromField(this.accessPath.apply(entity));
-        } catch (Exception e) {
-            throw Exceptions.handle()
-                            .to(OMA.LOG)
-                            .error(e)
-                            .withSystemErrorMessage(
-                                    "Unable to obtain a reference from entity ref field ('%s' in '%s'): %s (%s)",
-                                    getName(),
-                                    descriptor.getType().getName())
-                            .handle();
-        }
-    }
-
-    @Override
-    protected Object getValueFromField(Object target) {
-        return getRef(target).getKey();
-    }
-
-    @Override
-    public Object transformValue(Value value) {
-        return value.get();
-    }
-
-    @Override
-    protected Object transformToJDBC(Object object) {
-        return object;
-    }
-
-    @Override
-    protected Object transformFromJDBC(Value object) {
-        return object.get();
-    }
-
-    @Override
-    public void setValue(Object entity, Object object) {
-        this.setValueToField(object, entity);
-    }
-
-    @Override
-    protected void setValueToField(Object value, Object target) {
-        BlobSoftRef ref = getRef(target);
-        if (value == null || value instanceof Blob) {
-            ref.setBlob((Blob) value);
-        } else {
-            ref.setKey((String) value);
-        }
-    }
-
-    @Override
-    public void contributeToTable(Table table) {
-        table.getColumns().add(new TableColumn(this, Types.CHAR));
-    }
-
     @Override
     protected void determineLengths() {
-        BlobSoftRef templateReference = getRef(descriptor.getReferenceInstance());
+        BlobSoftRef templateReference = (BlobSoftRef) getRef(descriptor.getReferenceInstance());
         this.length =
                 templateReference.isSupportsURL() ? URL_COMPATIBLE_LENGTH : BlobHardRefProperty.DEFAULT_KEY_LENGTH;
     }
 
     @Override
     protected void onBeforeSaveChecks(Object entity) {
-        BlobSoftRef ref = getRef(entity);
         if (ref.changed && ref.isFilled() && ref.getBlob().isTemporary()) {
+        BlobHardRef ref = getRef(entity);
             throw Exceptions.handle()
                             .to(StorageUtils.LOG)
                             .withSystemErrorMessage(
